+++ modules/gov/utils_proposal.go
@@ -10,34 +10,33 @@ import (
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
-	tmctypes "github.com/tendermint/tendermint/rpc/core/types"
-
 	"google.golang.org/grpc/codes"
+
+	tmctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	"github.com/forbole/bdjuno/v3/types"
 
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 
-	inflationtypes "github.com/evmos/evmos/v6/x/inflation/types"
+	gov "github.com/cosmos/cosmos-sdk/x/gov/types"
+
+	inflationtypes "github.com/evmos/evmos/v10/x/inflation/types"
 )
 
 func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) error {
 	// Get the proposal
 	proposal, err := m.source.Proposal(height, id)
 	if err != nil {
-		// Check if proposal has reached the voting end time
-		passedVotingPeriod := blockTime.After(proposal.VotingEndTime)
-
-		if strings.Contains(err.Error(), codes.NotFound.String()) && passedVotingPeriod {
+		// Check if proposal exist on the chain
+		if strings.Contains(err.Error(), codes.NotFound.String()) && strings.Contains(err.Error(), "doesn't exist") {
 			// Handle case when a proposal is deleted from the chain (did not pass deposit period)
 			return m.updateDeletedProposalStatus(id)
 		}
 
 		return fmt.Errorf("error while getting proposal: %s", err)
 	}
-
 	err = m.updateProposalStatus(proposal)
 	if err != nil {
 		return fmt.Errorf("error while updating proposal status: %s", err)
@@ -52,7 +51,6 @@ func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) er
 	if err != nil {
 		return fmt.Errorf("error while updating account: %s", err)
 	}
-
 	err = m.handlePassedProposal(proposal, height)
 	if err != nil {
 		return fmt.Errorf("error while handling passed proposals: %s", err)
@@ -108,10 +106,10 @@ func (m *Module) handleParamChangeProposal(height int64, paramChangeProposal *pr
 			if err != nil {
 				return fmt.Errorf("error while updating ParamChangeProposal %s params : %s", distrtypes.ModuleName, err)
 			}
-		case govtypes.ModuleName:
+		case gov.ModuleName:
 			err = m.UpdateParams(height)
 			if err != nil {
-				return fmt.Errorf("error while updating ParamChangeProposal %s params : %s", govtypes.ModuleName, err)
+				return fmt.Errorf("error while updating ParamChangeProposal %s params : %s", gov.ModuleName, err)
 			}
 		case minttypes.ModuleName:
 			err = m.mintModule.UpdateParams(height)
@@ -140,7 +138,7 @@ func (m *Module) handleParamChangeProposal(height int64, paramChangeProposal *pr
 }
 
 // updateProposalStatus updates the given proposal status
-func (m *Module) updateProposalStatus(proposal govtypes.Proposal) error {
+func (m *Module) updateProposalStatus(proposal govtypesv1beta1.Proposal) error {
 	return m.db.UpdateProposal(
 		types.NewProposalUpdate(
 			proposal.ProposalId,
@@ -152,7 +150,7 @@ func (m *Module) updateProposalStatus(proposal govtypes.Proposal) error {
 }
 
 // updateProposalTallyResult updates the tally result associated with the given proposal
-func (m *Module) updateProposalTallyResult(proposal govtypes.Proposal) error {
+func (m *Module) updateProposalTallyResult(proposal govtypesv1beta1.Proposal) error {
 	height, err := m.db.GetLastBlockHeight()
 	if err != nil {
 		return err
@@ -166,17 +164,17 @@ func (m *Module) updateProposalTallyResult(proposal govtypes.Proposal) error {
 	return m.db.SaveTallyResults([]types.TallyResult{
 		types.NewTallyResult(
 			proposal.ProposalId,
-			result.Yes.String(),
-			result.Abstain.String(),
-			result.No.String(),
-			result.NoWithVeto.String(),
+			result.YesCount,
+			result.AbstainCount,
+			result.NoCount,
+			result.NoWithVetoCount,
 			height,
 		),
 	})
 }
 
 // updateAccounts updates any account that might be involved in the proposal (eg. fund community recipient)
-func (m *Module) updateAccounts(proposal govtypes.Proposal) error {
+func (m *Module) updateAccounts(proposal govtypesv1beta1.Proposal) error {
 	content, ok := proposal.Content.GetCachedValue().(*distrtypes.CommunityPoolSpendProposal)
 	if ok {
 		height, err := m.db.GetLastBlockHeight()
@@ -277,15 +275,15 @@ func findStatus(consAddr string, statuses []types.ValidatorStatus) (types.Valida
 	return types.ValidatorStatus{}, fmt.Errorf("cannot find status for validator with consensus address %s", consAddr)
 }
 
-func (m *Module) handlePassedProposal(proposal govtypes.Proposal, height int64) error {
-	if proposal.Status != govtypes.StatusPassed {
+func (m *Module) handlePassedProposal(proposal govtypesv1beta1.Proposal, height int64) error {
+	if proposal.Status != govtypesv1beta1.StatusPassed {
 		// If proposal status is not passed, do nothing
 		return nil
 	}
 
 	// Unpack proposal
-	var content govtypes.Content
-	err := m.db.EncodingConfig.Marshaler.UnpackAny(proposal.Content, &content)
+	var content govtypesv1beta1.Content
+	err := m.db.EncodingConfig.Codec.UnpackAny(proposal.Content, &content)
 	if err != nil {
 		return fmt.Errorf("error while handling ParamChangeProposal: %s", err)
 	}
@@ -297,14 +295,12 @@ func (m *Module) handlePassedProposal(proposal govtypes.Proposal, height int64) 
 		if err != nil {
 			return fmt.Errorf("error while updating params from ParamChangeProposal: %s", err)
 		}
-
 	case *upgradetypes.SoftwareUpgradeProposal:
 		// Store software upgrade plan while SoftwareUpgradeProposal passed
 		err = m.db.SaveSoftwareUpgradePlan(proposal.ProposalId, p.Plan, height)
 		if err != nil {
 			return fmt.Errorf("error while storing software upgrade plan: %s", err)
 		}
-
 	case *upgradetypes.CancelSoftwareUpgradeProposal:
 		// Delete software upgrade plan while CancelSoftwareUpgradeProposal passed
 		err = m.db.DeleteSoftwareUpgradePlan(proposal.ProposalId)