+++ modules/staking/source/remote/source_actions.go
@@ -3,7 +3,7 @@ package remote
 import (
 	"github.com/cosmos/cosmos-sdk/types/query"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/forbole/juno/v3/node/remote"
+	"github.com/forbole/juno/v4/node/remote"
 
 	"github.com/forbole/bdjuno/v3/utils"
 )